# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsApplicationsByStatus } from './adminStatsApplicationsByStatus';
import type { AdminStatsApplicationsByType } from './adminStatsApplicationsByType';
import type { AdminStatsUsersByRole } from './adminStatsUsersByRole';

export interface AdminStats {
  applicationsByStatus: AdminStatsApplicationsByStatus;
  applicationsByType: AdminStatsApplicationsByType;
  usersByRole: AdminStatsUsersByRole;
  newContactMessages: number;
  totalApplications: number;
  totalUsers: number;
}