# lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationDealType } from './applicationDealType';
import type { ApplicationStatus } from './applicationStatus';
import type { ApplicationType } from './applicationType';

export interface Application {
  id: string;
  userId: string;
  type: ApplicationType;
  name: string;
  email: string;
  company?: string | null;
  message: string;
  budget?: string | null;
  status: ApplicationStatus;
  dealType?: ApplicationDealType;
  createdAt: Date;
}