# lib/api-zod/src/generated/types/applicationNote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationNote {
  id: string;
  applicationId: string;
  authorId?: string | null;
  authorName?: string | null;
  body: string;
  createdAt: Date;
}