# lib/api-zod/src/generated/types/assistantChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantChatMessage } from './assistantChatMessage';

export interface AssistantChatInput {
  /** @minLength 1 */
  pass: string;
  /**
     * @minItems 1
     * @maxItems 20
     */
  messages: AssistantChatMessage[];
}