# lib/api-zod/src/generated/types/assistantChatMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantChatMessageRole } from './assistantChatMessageRole';

export interface AssistantChatMessage {
  role: AssistantChatMessageRole;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}