# lib/api-zod/src/generated/types/assistantVerifyOutput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantVerifyOutput {
  pass: string;
  expiresAt: Date;
}