# lib/api-zod/src/generated/types/createApplicationNoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApplicationNoteInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  body: string;
}