# lib/api-zod/src/generated/types/createContactMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactMessageInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  email: string;
  /** @maxLength 150 */
  subject?: string | null;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  message: string;
  /** Honeypot field; must be empty for human submissions. */
  website?: string | null;
}